refactor(jobs): split monolithic main.rs into modules
