+++ examples/jobs/src/config.rs
@@ -1,0 +1,11 @@
+use modo_db::DatabaseConfig;
+use serde::Deserialize;
+
+#[derive(Default, Deserialize)]
+pub(crate) struct Config {
+    #[serde(flatten)]
+    pub(crate) core: modo::config::AppConfig,
+    pub(crate) database: DatabaseConfig,
+    #[serde(default)]
+    pub(crate) jobs: modo_jobs::JobsConfig,
+}
